% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_transform.R
\name{coef_transform}
\alias{coef_transform}
\alias{coef_transform.dimred}
\title{Transform linear model coefficients to original feature space}
\usage{
\method{coef_transform}{dimred}(object, beta, alpha, ...)
}
\arguments{
\item{object}{Dimension reduction object.}

\item{beta}{Regression coefficients in the z-space (latent space).}

\item{alpha}{Intercept in the z-space (latent space).}

\item{...}{Currently ignored.}
}
\value{
A list with elements \code{beta} and \code{alpha} that give the model's 
regression coefficients and intercept, respectively, in the original feature space.
}
\description{
Transforms linear model coefficients from a model fitted using latent features z
back to the original feature space x. This is possible since the dimension reduction methods
PCA, SPCA and ISPCA are all linear, and consequently, if a linear model is fitted using 
features z, it will correspond exactly to a linear model with features x.
This function can be used to compute the regression coefficients and intercept of that model.
}
\examples{
\donttest{
### 

# load data
data('prostate', package = 'dimreduce')
x <- prostate$x
y <- prostate$y

# perform dimension reduction (use first two latent features)
dr <- ispca(x,y, nctot=2)
z <- predict(dr,x)

# fit model
model <- glm(y ~ z, family = binomial())
param_z <- coef(model)
alpha_z <- param_z[1]
beta_z <- param_z[2:3]

# transform the linear model back to original feature space
param_x <- coef_transform(dr, beta_z, alpha_z)
beta_x <- param_x$beta
alpha_x <- param_x$alpha

# plot the regression coefficients in the original space
plot(beta_x)

}


}
