% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colon}
\alias{colon}
\title{Colon cancer microarray data.}
\format{Binary classification data set.
\describe{
  \item{y}{target, 0 or 1.}
  \item{x}{features, 2000 in total.}
}}
\usage{
colon
}
\description{
Colon cancer microarray data.
}
\keyword{datasets}
