% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduce-package.R
\docType{package}
\name{dimreduce}
\alias{dimreduce}
\alias{dimreduce-package}
\title{Projection predictive feature selection}
\description{
Description

\pkg{dimreduce} is an R package to perform supervised dimension reduction.
}
\section{References}{


Bair, E., Hastie, T., Paul, D., and Tibshirani, R. (2006).
Prediction by supervised principal components. \emph{Journal
of the American Statistical Association}, 101(473):119-137.

Piironen, Juho and Vehtari, Aki (2018). Iterative supervised principal components.
To appear in \emph{Proceedings of the 21st International Conference on Artificial
Intelligence and Statistics (AISTATS)}.
}

