% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leukemia}
\alias{leukemia}
\title{Leukemia cancer microarray data.}
\format{Binary classification data set.
\describe{
  \item{y}{target, 0 or 1.}
  \item{x}{features, 7129 in total.}
}}
\usage{
leukemia
}
\description{
Leukemia cancer microarray data.
}
\keyword{datasets}
