% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ovarian}
\alias{ovarian}
\title{Ovarian cancer microarray data.}
\format{
Binary classification data set.
\describe{
  \item{y}{target, 0 or 1.}
  \item{x}{features, 1536 in total.}
}
}
\usage{
ovarian
}
\description{
Ovarian cancer microarray data.
}
\keyword{datasets}
