% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prostate}
\alias{prostate}
\title{Prostate cancer microarray data.}
\format{Binary classification data set.
\describe{
  \item{y}{target, 0 or 1.}
  \item{x}{features, 5966 in total.}
}}
\usage{
prostate
}
\description{
Prostate cancer microarray data.
}
\keyword{datasets}
