% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca.R
\name{spca}
\alias{spca}
\title{Supervised (and unsupervised) principal components}
\usage{
spca(
  x,
  y = NULL,
  nctot = NULL,
  ncsup = NULL,
  window = 500,
  exclude = NULL,
  verbose = TRUE,
  normalize = FALSE,
  preprocess = TRUE,
  alpha = NULL,
  perms = 1000,
  screenthresh = NULL,
  nfeat = NULL,
  sup.only = FALSE,
  ...
)
}
\arguments{
\item{x}{The original feature matrix, columns denoting the features and rows the instances.}

\item{y}{A vector with the observed target values we try to predict using \code{x}.
Can be factor for classification problems. If missing, then this function computes
standard unsupervised principal components.}

\item{nctot}{Total number of latent features to extract.}

\item{ncsup}{Maximum number of latent features to extract that use supervision.
If \code{nctot > ncsup}, then the remaining \code{nctot-ncsup} features are computed in
unsupervised manner (or ignored if \code{sup.only=TRUE}).}

\item{window}{Maximum number of features that will survive the screening and from which
the supervised components are computed. Affects also how the \code{screenthresh}-argument
is interpreted.}

\item{exclude}{Columns (variables) in \code{x} to ignore when extrating the new features.}

\item{verbose}{Whether to print some messages along the way.}

\item{normalize}{Whether to scale the extracted features so that they all have standard deviation
of one.}

\item{preprocess}{Whether to center and scale the features before extracting the new features.}

\item{alpha}{Significance level for the p-values of the univariate scores used to determine
which features survive the screening and are used to compute the supervised components.}

\item{perms}{Number of permutations to estimate the p-values for univariate scores.}

\item{screenthresh}{Value between 0 and 1 (or \code{NULL}). If not \code{NULL}, then no permutation tests are
run, and the supervised components
are computed among those features that have their univariate score equal or larger than this. 
Value 1 means that only the feature with the highest score survives the screening, whereas value 0 means that
the top \code{min(window, ncol(x))} survive the screening. Overwrites also \code{nfeat}-argument.}

\item{nfeat}{Number of features to retain in the screening step. If this option is used, then
the algorithm does not perform the permutation tests for the p-values, but instead computes the 
supervised components from those features that have their univariate score among the \code{nfeat} 
highest scores (in this case \code{perms} and \code{alpha} are ignored).}

\item{sup.only}{If \code{TRUE}, then no unsupervised components are ever computed even if the number of
supervised components that could be extracted was less than \code{nctot}.}

\item{...}{Currently ignored.}
}
\value{
spca-object that is similar to the object returned by \code{\link[stats]{prcomp}}.
The object will have the following elements:
\describe{
 \item{\code{w}}{The projection (or rotation) matrix W, that transforms the original data 
 \eqn{X} into the new features \eqn{Z = X W} .}
 \item{\code{z}}{The extracted latent features corresponding to the training inputs \eqn{X}.}
 \item{\code{v}}{Matrix \eqn{V} that is used to compute \eqn{W} when combining supervised and
 unsupervised components (see the Piironen and Vehtari (2018) for more information).}
 \item{\code{sdev}}{Standard deviations of the new features.}
 \item{\code{centers}}{Mean values for the original variables.}
 \item{\code{scales}}{Scales of the original variables.}
 \item{\code{exclude}}{Excluded variables.}
}
}
\description{
Computes dimension reduction based on the supervised principal components
algorithm. In essense, algorithm performs a screening step based on univariate
scores for the features, and then computes standard PCA on using only the retained
features.
}
\section{Details}{


In the original paper, the authors proposed estimating the screening threshold
using cross-validation for the model obtained when the extracted features are used
for regression or classification. This implementation performs the screening
based on the estimated p-values for the univariate scores (these are estimated using
a permutation test) and the screening step retains only those features with p-value
less than the specified level \code{alpha}.
}

\section{References}{


Bair, E., Hastie, T., Paul, D., and Tibshirani, R. (2006).
Prediction by supervised principal components. \emph{Journal
of the American Statistical Association}, 101(473):119-137.

Piironen, J. and Vehtari, A. (2018). Iterative supervised principal components.
In \emph{Proceedings of the 21st International Conference on Artificial
Intelligence and Statistics (AISTATS) PMLR 84: 106-114}.
}

\examples{
\donttest{
### 
dr <- spca(x,y, nctot=2)
z <- predict(dr, x) # the latent features
}

}
