\name{DimReduce-package}
\alias{DimReduce-package}
\alias{DimReduce}
\docType{package}
\title{
Non-linear Dimensionality Reduction techniques.
}
\description{
Non-linear Dimensionality reduction techniques that are applicable to data visualization, exploration, and mining. Includes standard spectral kernel methods and boxcox family stress functions for Multi-Dimensional scaling variants.
}
\details{
\tabular{ll}{
Package: \tab DimReduce\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-04\cr
License: \tab GPL\cr
Depends: \tab Rcpp (>= 0.10.3), RcppArmadillo (>= 0.2.0)\cr
}
This package provides an implementation of many spectral kernel methods, as well as Multidimensional Scaling via the Boxcox family of stress functions. Uses RcppArmadillo for efficient linear algebra.
}
\author{
Luke de Oliveira

Maintainer: Luke de Oliveira <luke.deoliveira@yale.edu>
}
% \references{
% ~~ Literature or other references for background information ~~
% }
% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
% ~~ the R documentation directory ~~
% \keyword{ package }
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% %% ~~ simple examples of the most important functions ~~
% }


% 
% move cancer to workshop

% B/M -- when?


