% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{featscores}
\alias{featscores}
\alias{featscore}
\alias{featscore.test}
\title{Univariate feature scores and significance tests}
\usage{
featscore(x, y, type = "pearson", exclude = NULL)

featscore.test(x, y, type = "pearson", exclude = NULL, test.max = FALSE,
  perms = 1000)
}
\arguments{
\item{x}{The original feature matrix, columns denoting the features and rows the instances.}

\item{y}{A vector with the observed target values we try to predict using \code{x}.
Can be factor for classification problems.}

\item{type}{Score type. One of 'pearson', 'kendall', 'spearman', or 'runs'. The first three
denote the type of correlation computed (will be passed to \link[stats]{cor}), whereas runs
denote the runs test, that can potentially detect any nonlinear relationship.}

\item{exclude}{Columns (variables) in x to ignore. The score will be zero for these.}

\item{test.max}{If TRUE, compute the p-value for the maximum of the univariate scores.
If FALSE (default), compute p-values separately for each feature.}

\item{perms}{Number of random permutations to estimate the p-values for univariate scores.}

\item{...}{Currently ignored.}
}
\value{
A vector giving the univariate scores for each feature.
}
\description{
\code{featscore} computes the univariate scores for each feature, and 
\code{featscore.test} can be used to compute the corresponding p-values to measure 
statistical significance of these values.
}
\section{Details}{


Univariate scores are a useful technique to assess variable relevances, and 
can be used for screening. The paper below has nice discussion and practical
tips for how to use univariate scores and when they are appropriate.
}

\section{References}{


Neal, R. and Zhang, J. (2006). High dimensional classification with Bayesian 
neural networks and Dirichlet diffusion trees. 
In Guyon, I., Gunn, S., Nikravesh, M., and Zadeh, L. A., editors, 
\emph{Feature Extraction, Foundations and Applications}, pages 265-296. Springer.
}

\examples{
\donttest{
### 
s <- featscore(x,y) # correlation
s <- featscore(x,y, type='runs') # runs test, can detect nonlinear effects
s <- featscore.test(x,y,type='spearman') # p-values for the rank-correlations
}

}
