% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ispca.R
\name{ispca}
\alias{ispca}
\title{Iterative supervised principal components}
\usage{
ispca(x, y, thresh = NULL, nthresh = NULL, ncsup = NULL, nctot = NULL,
  exclude = NULL, window = 500, verbose = TRUE, min_score = 1e-04,
  normalize = TRUE, center = TRUE, scale = TRUE, permtest = TRUE,
  permtest_type = "max-marginal", perms = 1000, alpha = 0.01,
  method = "svd", ...)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{...}{Currently ignored.}
}
\value{
ispca-object that is similar in spirit to the object returned by \code{\link[stats]{prcomp}}.
The object will have the following elements:
\describe{
 \item{\code{rotation}}{ The projection (or rotation) matrix W, that transforms the original data 
 \eqn{X} into the new features \eqn{Z = X W} .}
 \item{\code{x}}{TODO.}
 \item{\code{v}}{Matrix \eqn{V} that is used to compute \eqn{W}. The columns of \eqn{V} indicate
 which variables become active at each iteration (see the paper below for more information).}
 \item{\code{sdev}}{Standard deviations of the new features.}
 \item{\code{ncsup}}{How many supervised components was extracted (the rest are computed 
 in an unsupervised manner).}
 \item{\code{centers}}{Mean values for the original variables.}
 \item{\code{scales}}{Scales of the original variables.}
 \item{\code{exclude}}{Excluded variables.}
}
}
\description{
Computes dimension reduction based on the iterative supervised principal components
algorithm.
}
\section{References}{


Piironen, Juho and Vehtari, Aki (2018). Iterative supervised principal components.
To appear in \emph{Proceedings of the 21st International Conference on Artificial
Intelligence and Statistics (AISTATS)}.
}

\examples{
\donttest{
### 
dr <- ispca(x,y, nctot=10)
}

}
