% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca.R
\name{spca}
\alias{spca}
\title{Supervised (and unsupervised) principal components}
\usage{
spca(x, y = NULL, family = gaussian(), ncsup = NULL, nctot = NULL,
  exclude = NULL, verbose = TRUE, normalize = TRUE, preprocess = TRUE,
  alpha = NULL, perms = 1000, ...)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{...}{Currently ignored.}
}
\value{
spca-object that is similar to the object returned by \code{\link[stats]{prcomp}}.
The object will have the following elements:
\describe{
 \item{\code{foo}}{TODO.}
 \item{\code{bar}}{TODO.}
}
}
\description{
Computes dimension reduction based on the supervised principal components
algorithm. In essense, algorithm performs a screening step based on univariate
scores for the features, and then computes standard PCA on using only the retained
features.
}
\section{Details}{


In the original paper, the authors proposed estimating the screening threshold
using cross-validation for the model obtained when the extracted features are used
for regression or classification. This implementation performs the screening
based on the estimated p-values for the univariate scores (these are estimated using
a permutation test) and the screening step retains only those features with p-value
less than the specified level \code{alpha}.
}

\section{References}{


Bair, E., Hastie, T., Paul, D., and Tibshirani, R. (2006).
Prediction by supervised principal components. \emph{Journal
of the American Statistical Association}, 101(473):119-137.
}

\examples{
\donttest{
### 
dr <- spca(x,y, nctot=10)
}

}
